Rastreabilidade e logs nas APIs

+++ microservice-product-api/src/main/java/br/com/rrssistema/microserviceproductapi/modules/product/service/ProductService.java
@@ -8,9 +8,11 @@ import br.com.rrssistema.microserviceproductapi.modules.product.model.Product;
 import br.com.rrssistema.microserviceproductapi.modules.product.repository.ProductRepository;
 import br.com.rrssistema.microserviceproductapi.modules.sales.client.SalesClient;
 import br.com.rrssistema.microserviceproductapi.modules.sales.dto.SalesConfirmationDTO;
+import br.com.rrssistema.microserviceproductapi.modules.sales.dto.SalesProductResponse;
 import br.com.rrssistema.microserviceproductapi.modules.sales.enums.SalesStatus;
 import br.com.rrssistema.microserviceproductapi.modules.sales.rabbitmq.SalesConfirmationSender;
 import br.com.rrssistema.microserviceproductapi.modules.supplier.service.SupplierService;
+import com.fasterxml.jackson.databind.ObjectMapper;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
@@ -20,6 +22,7 @@ import java.util.ArrayList;
 import java.util.List;
 import java.util.stream.Collectors;
 
+import static br.com.rrssistema.microserviceproductapi.config.RequestUtil.getCurrentRequest;
 import static org.springframework.util.ObjectUtils.isEmpty;
 
 @Slf4j
@@ -27,6 +30,8 @@ import static org.springframework.util.ObjectUtils.isEmpty;
 public class ProductService {
 
     private static final  Integer ZERO = 0;
+    private static final String TRANSACTION_ID = "transactionid";
+    private static final String SERVICE_ID = "serviceid";
     @Autowired
     private ProductRepository productRepository;
     @Autowired
@@ -118,8 +123,15 @@ public class ProductService {
 
     public SuccessResponse delete(Integer id) {
         validateInformedId(id);
+        if (!productRepository.existsById(id)) {
+            throw new ValidationException("The product does not exists.");
+        }
+        var sales = getSalesByProductId(id);
+        if (!isEmpty(sales.getSalesIds())) {
+            throw new ValidationException("The product cannot be deleted. There are sales for it.");
+        }
         productRepository.deleteById(id);
-        return SuccessResponse.create("The product ID must be informed");
+        return SuccessResponse.create("The product was deleted.");
     }
 
     public Boolean existsByCategoryId(Integer categoryId) {
@@ -163,7 +175,8 @@ public class ProductService {
             updateStock(product);
         } catch (Exception ex) {
             log.error("Error while trying to update stock for message with error: {}", ex.getMessage(), ex);
-            var rejectedMessage = new SalesConfirmationDTO(product.getSalesId(), SalesStatus.REJECTED);
+            var rejectedMessage = new SalesConfirmationDTO(product.getSalesId(),
+                    SalesStatus.REJECTED, product.getTransactionid());
             salesConfirmationSender.sendSalesConfirmationMessage(rejectedMessage);
         }
     }
@@ -196,7 +209,8 @@ public class ProductService {
                 });
         if(!isEmpty(productRepository)) {
             productRepository.saveAll(productsForUpdate);
-            var approvedMessage = new SalesConfirmationDTO(product.getSalesId(), SalesStatus.APPROVED);
+            var approvedMessage = new SalesConfirmationDTO(product.getSalesId(),
+                    SalesStatus.APPROVED, product.getTransactionid());
             salesConfirmationSender.sendSalesConfirmationMessage(approvedMessage);
         }
     }
@@ -210,24 +224,48 @@ public class ProductService {
 
     public ProductSalesResponse findProductSales(Integer id) {
         var product = findById(id);
+        var sales = getSalesByProductId(product.getId());
+        return ProductSalesResponse.of(product, sales.getSalesIds());
+    }
+
+    private SalesProductResponse getSalesByProductId(Integer productId) {
         try {
-            var sales = salesClient
-                    .findSalesByProductId(product.getId())
+            var currentRequest = getCurrentRequest();
+            var transactionId = currentRequest.getHeader(TRANSACTION_ID);
+            var serviceId = currentRequest.getAttribute(SERVICE_ID);
+            log.info("Sending GET request to orders by productId with data {} | [transactionID: {} | serviceID: {}]",
+                    productId, transactionId, serviceId);
+            var response = salesClient
+                    .findSalesByProductId(productId)
                     .orElseThrow(() -> new ValidationException("The sales was not found by this product."));
-            return ProductSalesResponse.of(product, sales.getSalesIds());
+            log.info("Receiving response from orders by productId with data {} | [transactionID: {} | serviceID: {}]",
+                    new ObjectMapper().writeValueAsString(response), transactionId, serviceId);
+            return response;
         } catch (Exception ex) {
-            throw new ValidationException("There was an error trying to get the product's sales.");
+            throw new ValidationException("The sales could not be found.");
         }
     }
 
     public SuccessResponse checkProductsStock(ProductCheckStockRequest request) {
-        if(isEmpty(request) || isEmpty(request.getProducts())) {
-            throw new ValidationException("The request data and products must be informed");
+        try {
+            var currentRequest = getCurrentRequest();
+            var transactionId = currentRequest.getHeader(TRANSACTION_ID);
+            var serviceId = currentRequest.getAttribute(SERVICE_ID);
+            log.info("Request to POST product stock with data {} | [transactionID: {} | serviceID: {}]",
+                    new ObjectMapper().writeValueAsString(request), transactionId, serviceId);
+            if (isEmpty(request) || isEmpty(request.getProducts())) {
+                throw new ValidationException("The request data and products must be informed.");
+            }
+            request
+                    .getProducts()
+                    .forEach(this::validateStock);
+            var response = SuccessResponse.create("The stock is ok!");
+            log.info("Response to POST product stock with data {} | [transactionID: {} | serviceID: {}]",
+                    new ObjectMapper().writeValueAsString(response), transactionId, serviceId);
+            return response;
+        } catch (Exception ex) {
+            throw new ValidationException(ex.getMessage());
         }
-        request
-                .getProducts()
-                .forEach(this::validateStock);
-        return SuccessResponse.create("The stock is ok!");
     }
 
     private void validateStock(ProductQuantityDTO productQuantity) {
